requirement.txt
streamlit
flask
google-generativeai
networkx
pyvis
python-dotenv
requests
beautifulsoup4
PyMuPDF
gunicorn